// src/app/add-thesis-dialog/thesis-create-dto.ts
export interface ThesisCreateDTO {
  studentName: string | null;
  levelOfStudies: string | null;
  curriculum: string | null;
  languageOfThesis: string | null;
  volumeEcts: number | null;
  titleEstonian: string | null;
  titleEnglish: string | null;
  contextOfResearch: string | null;
  studentContribution: string | null;
  strengthOfThesis: string | null;
  limitationOfThesis: string | null;
  cooperation: string | null;
  additionalComments: string | null;
}